Update huffmann-tree.py
I initialized the method self.added_texts, which is to be accessed in reset_program(), in the list self.added_texts in the HuffmanTreeApp class (in the constructor). Then the numbers are also deleted alongside the tree.
Furthermore, I made sure that no error occurs if the user accidentally clicks anywhere on the empty canvas (in the connect_mode() function).